shared_dir/requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.0.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async==0.6.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bidict==0.22.1
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.1
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
charset-normalizer==3.3.0
click==8.1.7
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
contourpy==1.1.1
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.11.0
Cython==3.0.2
decorator==4.4.2
dnspython==2.1.0
dominate==2.8.0
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-socketio==0.0.10
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
Flask==3.0.0
fonttools==4.42.1
frozenlist==1.4.0
future==0.17.1
futures==3.0.5
gast==0.2.2
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h11==0.14.0
h5py==2.9.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython @ file:///tmp/build/80754af9/ipython_1610724799192/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
Markdown==3.4.4
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.7.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==3.0.5
multidict==6.0.4
networkx==3.1
nicegui==1.3.15
numpy==1.20.3
oauthlib==3.2.2
olefile==0.46
opencv-contrib-python==4.4.0.46
opencv-python==4.8.0.76
OpenEXR==1.3.9
orjson==3.9.7
packaging==23.1
pandas==2.0.3
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786786540/work
pkginfo==1.7.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==4.24.3
pscript==0.7.7
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-etcd==0.4.5
python-multipart==0.0.6
python-socketio==5.9.0
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyWavelets==1.4.1
PyYAML==5.3.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.1
retinaface-pytorch==0.0.8
rsa==4.9
ruamel_yaml==0.15.87
scikit-image==0.16.2
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.12.2
simple-websocket==0.10.1
six @ file:///tmp/build/80754af9/six_1605205327372/work
sniffio==1.3.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
starlette==0.27.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==1.9.0
torchelastic==0.2.1
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.3.3
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typing_extensions==4.8.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
uvicorn==0.22.0
uvloop==0.17.0
vbuild==0.8.2
visdom==0.2.4
watchfiles==0.20.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==3.0.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
